clean up code and address comments
