+++ src/java/org/apache/cassandra/locator/InetAddressAndPort.java
@@ -69,8 +69,6 @@ public final class InetAddressAndPort extends InetSocketAddress implements Compa
 
     public final byte[] addressBytes;
 
-    public static final long IPV6_SIZE = ObjectSizes.measureDeep(new InetSocketAddress(getIpvAddress(16), 42));
-
     @VisibleForTesting
     InetAddressAndPort(InetAddress address, byte[] addressBytes, int port)
     {
@@ -150,23 +148,6 @@ public final class InetAddressAndPort extends InetSocketAddress implements Compa
         {
             return address.getAddress().getHostAddress();
         }
-    }
-
-    public static InetAddress getIpvAddress(int size)
-    {
-        if (size == 16 || size ==4)
-        {
-            try
-            {
-                return InetAddress.getByAddress(new byte[size]);
-            }
-            catch (UnknownHostException e)
-            {
-                e.printStackTrace();
-                throw new IllegalArgumentException("Invalid size of a byte array when getting and ipv address: " + size);
-            }
-        }
-        else throw new IllegalArgumentException("Excpected a byte array size of 4 or 16 for an ipv address but got: " + size);
     }
 
     @Override
