+++ src/java/org/apache/cassandra/utils/TimeUUID.java
@@ -82,6 +82,7 @@ public class TimeUUID implements Serializable, Comparable<TimeUUID>
     private static final long MIN_CLOCK_SEQ_AND_NODE = 0x8080808080808080L;
     private static final long MAX_CLOCK_SEQ_AND_NODE = 0x7f7f7f7f7f7f7f7fL;
 
+    public static final long TIMESTAMP_SIZE = ObjectSizes.measureDeep(new TimeUUID(10,10));
 
     final long uuidTimestamp, lsb;
